Extract MovecraftCombatHook
