+++ cannons-bukkit/src/main/resources/config.yml
@@ -19,7 +19,8 @@ hooks:
     enabled: true
   movecraft:
     enabled: true
-    damageTracking: true
+  movecraftCombat:
+    enabled: true
 
 
 cannonLimits:
